feat: review changes
- moved StarkHash from core to pedersen
- StarkHash is now copy
- removed hex parsing functionality
- added overflow check to `from_be_bytes`
- moved pedersen field + curve down a level
