+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod config;
-pub mod core;
 pub mod ethereum;
 pub mod pedersen;
 pub mod rpc;
