+++ src/pedersen.rs
@@ -1,226 +1,41 @@
-use bitvec::{array::BitArray, order::Lsb0, slice::BitSlice};
-use ff::{Field, PrimeField};
+use crate::pedersen::field::{
+    FieldElement, FieldElementRepr, PEDERSEN_P0, PEDERSEN_P1, PEDERSEN_P2, PEDERSEN_P3, PEDERSEN_P4,
+};
 
-use crate::core::StarkHash;
+use ff::PrimeField;
 
-/// The field primitive used by [PedersenHash]
-#[derive(PrimeField)]
-#[PrimeFieldModulus = "3618502788666131213697322783095070105623107215331596699973092056135872020481"]
-#[PrimeFieldGenerator = "7"]
-#[PrimeFieldReprEndianness = "big"]
-struct FieldElement([u64; 4]);
+mod field;
+/// The main hash code used by Starknet.
+///
+/// Contains 251 bits of data and is generated by the [pedersen_hash] function.
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct StarkHash([u8; 32]);
 
-impl From<StarkHash> for FieldElement {
-    fn from(hash: StarkHash) -> Self {
-        debug_assert_eq!(
-            std::mem::size_of::<FieldElement>(),
-            std::mem::size_of::<StarkHash>()
-        );
-        Self::from_repr(FieldElementRepr(hash.to_be_bytes())).unwrap()
+/// Error returned by [StarkHash::from_be_bytes] indicating that
+/// more than the allowed 251 bits were set.
+#[derive(Debug, PartialEq, Clone, Copy)]
+pub struct OverflowError;
+
+impl StarkHash {
+    /// Returns the big-endian representation of this [StarkHash].
+    pub fn to_be_bytes(self) -> [u8; 32] {
+        self.0
     }
-}
 
-impl From<FieldElement> for StarkHash {
-    fn from(fp: FieldElement) -> Self {
-        debug_assert_eq!(
-            std::mem::size_of::<FieldElement>(),
-            std::mem::size_of::<StarkHash>()
-        );
-        StarkHash::from_be_bytes(fp.to_repr().0)
-    }
-}
-
-impl FieldElement {
-    /// Transforms [FieldElement] into little endian bit representation.
-    fn into_bits(mut self) -> BitArray<Lsb0, [u64; 4]> {
-        #[cfg(not(target_endian = "little"))]
-        {
-            todo!("untested and probably unimplemented: big-endian targets")
-        }
-
-        #[cfg(target_endian = "little")]
-        {
-            self.mont_reduce(
-                self.0[0usize],
-                self.0[1usize],
-                self.0[2usize],
-                self.0[3usize],
-                0,
-                0,
-                0,
-                0,
-            );
-
-            self.0.into()
+    /// Creates a [StarkHash] from big-endian bytes.
+    ///
+    /// Returns [OverflowError] if more than 251 bits are set.
+    pub fn from_be_bytes(bytes: [u8; 32]) -> Result<Self, OverflowError> {
+        match bytes[0] {
+            0..=0b0000_0111 => Ok(Self(bytes)),
+            _ => Err(OverflowError),
         }
     }
 }
 
-/// A point on an elliptic curve over [FieldElement].
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct CurvePoint {
-    x: FieldElement,
-    y: FieldElement,
-    infinity: bool,
-}
-
-impl CurvePoint {
-    fn identity() -> CurvePoint {
-        Self {
-            x: FieldElement::zero(),
-            y: FieldElement::zero(),
-            infinity: true,
-        }
-    }
-
-    fn double(&self) -> CurvePoint {
-        if self.infinity {
-            return self.clone();
-        }
-
-        // l = (3x^2+a)/2y with a=1 from stark curve
-        let lambda = {
-            let two = FieldElement::one() + FieldElement::one();
-            let three = two + FieldElement::one();
-            let dividend = three * (self.x * self.x) + FieldElement::one();
-            let divisor_inv = (two * self.y).invert().unwrap();
-            dividend * divisor_inv
-        };
-
-        let result_x = (lambda * lambda) - self.x - self.x;
-        let result_y = lambda * (self.x - result_x) - self.y;
-
-        CurvePoint {
-            x: result_x,
-            y: result_y,
-            infinity: false,
-        }
-    }
-
-    fn add(&self, other: &CurvePoint) -> CurvePoint {
-        if self.infinity {
-            return other.clone();
-        }
-        if other.infinity {
-            return self.clone();
-        }
-
-        // l = (y2-y1)/(x2-x1)
-        let lambda = {
-            let dividend = other.y - self.y;
-            let divisor_inv = (other.x - self.x).invert().unwrap();
-            dividend * divisor_inv
-        };
-
-        let result_x = (lambda * lambda) - self.x - other.x;
-        let result_y = lambda * (self.x - result_x) - self.y;
-
-        CurvePoint {
-            x: result_x,
-            y: result_y,
-            infinity: false,
-        }
-    }
-
-    fn multiply(&self, bits: &BitSlice<Lsb0, u64>) -> CurvePoint {
-        let mut product = CurvePoint::identity();
-        for b in bits.iter().rev() {
-            product = product.double();
-            if *b {
-                product = product.add(self);
-            }
-        }
-
-        product
-    }
-}
-
-/// Montgomery representation of the Stark curve constant P0.
-const PEDERSEN_P0: CurvePoint = CurvePoint {
-    x: FieldElement([
-        1933903796324928314,
-        7739989395386261137,
-        1641324389046377921,
-        316327189671755572,
-    ]),
-    y: FieldElement([
-        14252083571674603243,
-        12587053260418384210,
-        4798858472748676776,
-        81375596133053150,
-    ]),
-    infinity: false,
-};
-
-/// Montgomery representation of the Stark curve constant P1.
-const PEDERSEN_P1: CurvePoint = CurvePoint {
-    x: FieldElement([
-        3602345268353203007,
-        13758484295849329960,
-        518715844721862878,
-        241691544791834578,
-    ]),
-    y: FieldElement([
-        13441546676070136227,
-        13001553326386915570,
-        433857700841878496,
-        368891789801938570,
-    ]),
-    infinity: false,
-};
-
-/// Montgomery representation of the Stark curve constant P2.
-const PEDERSEN_P2: CurvePoint = CurvePoint {
-    x: FieldElement([
-        16491878934996302286,
-        12382025591154462459,
-        10043949394709899044,
-        253000153565733272,
-    ]),
-    y: FieldElement([
-        13950428914333633429,
-        2545498000137298346,
-        5191292837124484988,
-        285630633187035523,
-    ]),
-    infinity: false,
-};
-
-/// Montgomery representation of the Stark curve constant P3.
-const PEDERSEN_P3: CurvePoint = CurvePoint {
-    x: FieldElement([
-        1203723169299412240,
-        18195981508842736832,
-        12916675983929588442,
-        338510149841406402,
-    ]),
-    y: FieldElement([
-        12352616181161700245,
-        11743524503750604092,
-        11088962269971685343,
-        161068411212710156,
-    ]),
-    infinity: false,
-};
-
-/// Montgomery representation of the Stark curve constant P4.
-const PEDERSEN_P4: CurvePoint = CurvePoint {
-    x: FieldElement([
-        1145636535101238356,
-        10664803185694787051,
-        299781701614706065,
-        425493972656615276,
-    ]),
-    y: FieldElement([
-        8187986478389849302,
-        4428713245976508844,
-        6033691581221864148,
-        345457391846365716,
-    ]),
-    infinity: false,
-};
-
-/// Computes the Starknet Pedersen hash on `a` and `b`.
+/// Computes the [Starknet Pedersen hash] on `a` and `b`.
+///
+/// [Starknet Pedersen hash]: https://docs.starkware.co/starkex-v3/crypto/pedersen-hash-function
 pub fn pedersen_hash(a: StarkHash, b: StarkHash) -> StarkHash {
     let mut result = PEDERSEN_P0.clone();
     let a = FieldElement::from(a).into_bits();
@@ -246,166 +61,48 @@ pub fn pedersen_hash(a: StarkHash, b: StarkHash) -> StarkHash {
     StarkHash::from(result.x)
 }
 
+impl From<StarkHash> for FieldElement {
+    fn from(hash: StarkHash) -> Self {
+        debug_assert_eq!(
+            std::mem::size_of::<FieldElement>(),
+            std::mem::size_of::<StarkHash>()
+        );
+        Self::from_repr(FieldElementRepr(hash.to_be_bytes())).unwrap()
+    }
+}
+
+impl From<FieldElement> for StarkHash {
+    fn from(fp: FieldElement) -> Self {
+        debug_assert_eq!(
+            std::mem::size_of::<FieldElement>(),
+            std::mem::size_of::<StarkHash>()
+        );
+        // unwrap is safe because the FieldElement and StarkHash
+        // should both be 251 bits only.
+        StarkHash::from_be_bytes(fp.to_repr().0).unwrap()
+    }
+}
 #[cfg(test)]
 mod tests {
     use super::*;
     use pretty_assertions::assert_eq;
 
-    mod to_le_bits_rev {
-        use super::*;
-        use pretty_assertions::assert_eq;
-
-        #[test]
-        fn zero() {
-            let zero = FieldElement::zero().into_bits();
-            let expected = BitArray::<Lsb0, [u64; 4]>::default();
-
-            assert_eq!(zero, expected);
-        }
-
-        #[test]
-        fn one() {
-            let one = FieldElement::one().into_bits();
-
-            let mut expected = BitArray::<Lsb0, [u64; 4]>::default();
-            expected.set(0, true);
-
-            assert_eq!(one, expected);
-        }
-
-        #[test]
-        fn two() {
-            let two = (FieldElement::one() + FieldElement::one()).into_bits();
-
-            let mut expected = BitArray::<Lsb0, [u64; 4]>::default();
-            expected.set(1, true);
-
-            assert_eq!(two, expected);
-        }
-    }
-
-    mod curve {
-        use super::*;
-        use pretty_assertions::assert_eq;
-
-        fn curve_from_xy_str(x: &str, y: &str) -> CurvePoint {
-            let x = FieldElement::from_str_vartime(x).expect("Curve x-value invalid");
-            let y = FieldElement::from_str_vartime(y).expect("Curve y-value invalid");
-            CurvePoint {
-                x,
-                y,
-                infinity: false,
-            }
-        }
-
-        fn curve_generator() -> CurvePoint {
-            curve_from_xy_str(
-                "874739451078007766457464989774322083649278607533249481151382481072868806602",
-                "152666792071518830868575557812948353041420400780739481342941381225525861407",
-            )
-        }
-
-        #[test]
-        fn double() {
-            let g_double = curve_generator().double();
-            let expected = curve_from_xy_str(
-                "3324833730090626974525872402899302150520188025637965566623476530814354734325",
-                "3147007486456030910661996439995670279305852583596209647900952752170983517249",
-            );
-            assert_eq!(g_double, expected);
-        }
-
-        #[test]
-        fn double_and_add() {
-            let g = curve_generator();
-            let g_double = g.double();
-            let g_triple = g_double.add(&g);
-            let expected = curve_from_xy_str(
-                "1839793652349538280924927302501143912227271479439798783640887258675143576352",
-                "3564972295958783757568195431080951091358810058262272733141798511604612925062",
-            );
-            assert_eq!(g_triple, expected);
-        }
-
-        #[test]
-        fn multiply() {
-            let three =
-                (FieldElement::one() + FieldElement::one() + FieldElement::one()).into_bits();
-            let g = curve_generator();
-            let g_triple = g.multiply(&three);
-            let expected = curve_from_xy_str(
-                "1839793652349538280924927302501143912227271479439798783640887258675143576352",
-                "3564972295958783757568195431080951091358810058262272733141798511604612925062",
-            );
-            assert_eq!(g_triple, expected);
-        }
-
-        #[test]
-        fn p0() {
-            let expected = curve_from_xy_str(
-                "2089986280348253421170679821480865132823066470938446095505822317253594081284",
-                "1713931329540660377023406109199410414810705867260802078187082345529207694986",
-            );
-
-            assert_eq!(PEDERSEN_P0, expected);
-        }
-
-        #[test]
-        fn p1() {
-            let expected = curve_from_xy_str(
-                "996781205833008774514500082376783249102396023663454813447423147977397232763",
-                "1668503676786377725805489344771023921079126552019160156920634619255970485781",
-            );
-
-            assert_eq!(PEDERSEN_P1, expected);
-        }
-
-        #[test]
-        fn p2() {
-            let expected = curve_from_xy_str(
-                "2251563274489750535117886426533222435294046428347329203627021249169616184184",
-                "1798716007562728905295480679789526322175868328062420237419143593021674992973",
-            );
-
-            assert_eq!(PEDERSEN_P2, expected);
-        }
-
-        #[test]
-        fn p3() {
-            let expected = curve_from_xy_str(
-                "2138414695194151160943305727036575959195309218611738193261179310511854807447",
-                "113410276730064486255102093846540133784865286929052426931474106396135072156",
-            );
-
-            assert_eq!(PEDERSEN_P3, expected);
-        }
-
-        #[test]
-        fn p4() {
-            let expected = curve_from_xy_str(
-                "2379962749567351885752724891227938183011949129833673362440656643086021394946",
-                "776496453633298175483985398648758586525933812536653089401905292063708816422",
-            );
-
-            assert_eq!(PEDERSEN_P4, expected);
-        }
-    }
-
     #[test]
     fn hash() {
-        // Test vector from https://github.com/starkware-libs/crypto-cpp/blob/master/src/starkware/crypto/pedersen_hash_test.cc
-        let a = StarkHash::from_hex_str(
-            "0x3d937c035c878245caf64531a5756109c53068da139362728feb561405371cb",
-        )
-        .unwrap();
-        let b = StarkHash::from_hex_str(
-            "0x208a0a10250e382e1e4bbe2880906c2791bf6275695e02fbbc6aeff9cd8b31a",
-        )
-        .unwrap();
-        let expected = StarkHash::from_hex_str(
-            "0x30e480bed5fe53fa909cc0f8c4d99b8f9f2c016be4c41e13a4848797979c662",
-        )
-        .unwrap();
+        // Test vectors from https://github.com/starkware-libs/crypto-cpp/blob/master/src/starkware/crypto/pedersen_hash_test.cc
+        let a = "03d937c035c878245caf64531a5756109c53068da139362728feb561405371cb";
+        let b = "0208a0a10250e382e1e4bbe2880906c2791bf6275695e02fbbc6aeff9cd8b31a";
+        let expected = "030e480bed5fe53fa909cc0f8c4d99b8f9f2c016be4c41e13a4848797979c662";
+
+        fn parse_hex(str: &str) -> [u8; 32] {
+            let mut buf = [0; 32];
+            hex::decode_to_slice(str, &mut buf).unwrap();
+            buf
+        }
+
+        let a = StarkHash::from_be_bytes(parse_hex(a)).unwrap();
+        let b = StarkHash::from_be_bytes(parse_hex(b)).unwrap();
+        let expected = StarkHash::from_be_bytes(parse_hex(expected)).unwrap();
 
         let hash = pedersen_hash(a, b);
 
@@ -413,32 +110,35 @@ mod tests {
     }
 
     #[test]
-    fn hash_to_field_element() {
-        // Compare decimal field point and hash hex string constructions.
-        let fp = FieldElement::from_str_vartime(
-            "1740729136829561885683894917751815192814966525555656371386868611731128807883",
-        )
-        .unwrap();
-        let hash = StarkHash::from_hex_str(
-            "0x3d937c035c878245caf64531a5756109c53068da139362728feb561405371cb",
-        )
-        .unwrap();
-
-        assert_eq!(FieldElement::from(hash), fp);
+    fn bytes_round_trip() {
+        let original = [
+            0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D,
+            0x0E, 0x0F, 0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1A, 0x1B,
+            0x1C, 0x1D, 0x1E, 0x1F,
+        ];
+        let hash = StarkHash::from_be_bytes(original).unwrap();
+        let bytes = hash.to_be_bytes();
+        assert_eq!(bytes, original);
     }
 
     #[test]
-    fn field_element_to_hash() {
-        // Compare decimal field point and hash hex string constructions.
-        let fp = FieldElement::from_str_vartime(
-            "1740729136829561885683894917751815192814966525555656371386868611731128807883",
-        )
-        .unwrap();
-        let hash = StarkHash::from_hex_str(
-            "0x3d937c035c878245caf64531a5756109c53068da139362728feb561405371cb",
-        )
-        .unwrap();
+    fn from_bytes_overflow() {
+        // Set the 252nd bit (which is invalid).
+        let mut bytes = [0; 32];
+        bytes[0] = 0b0000_1000;
+        assert_eq!(StarkHash::from_be_bytes(bytes), Err(OverflowError));
+    }
 
-        assert_eq!(StarkHash::from(fp), hash);
+    #[test]
+    fn hash_field_round_trip() {
+        let bytes = [
+            0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D,
+            0x0E, 0x0F, 0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1A, 0x1B,
+            0x1C, 0x1D, 0x1E, 0x1F,
+        ];
+        let original = StarkHash::from_be_bytes(bytes).unwrap();
+        let fp = FieldElement::from(original);
+        let hash = StarkHash::from(fp);
+        assert_eq!(hash, original);
     }
 }
